power6_backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.23.2
httpx==0.27.0
httptools==0.5.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9

# Optional tools & CLI
click==8.1.7
typer==0.9.0
colorama==0.4.6
packaging==23.2

# Data/Parsing
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1
pyyaml==6.0

# Background / scheduling
anyio==4.3.0

# SQLite + PostgreSQL & Testing
sqlalchemy==2.0.41
sqlmodel==0.0.16
pytest==8.2.2

# Utilities and core deps
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
h11==0.16.0
pyasn1==0.6.1
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
urllib3==2.5.0
wheel==0.45.1

# Stripe integration
stripe==12.4.0

# Web server
starlette==0.35.1

# Crypto, Security, Serialization
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.5
ecdsa==0.19.1

# Testing / Dev tools
iniconfig==2.1.0
pluggy==1.6.0

# Additional deps
annotated-types==0.7.0
greenlet==3.2.3
pycparser==2.22
